ISPN-891 Generated schema uses xs:sequence when not required
